Abstract code, begin adding functionality to login button
Created a login view model for code organization. Also printing after login button hit to test functionality, no real purpose yet. Also trims whitespace on email and password inpus. #10
