+++ Sweat Social/App/Sweat_SocialApp.swift
@@ -6,7 +6,7 @@
 //
 
 import SwiftUI
-import Firebase
+import FirebaseCore
 
 @main
 struct Sweat_SocialApp: App {
