+++ Sweat Social/ViewModels/LoginViewModel.swift
@@ -6,6 +6,21 @@
 //
 
 import Foundation
-import Firebase
+
+class LoginViewViewModel: ObservableObject {
+    @Published var email = ""
+    @Published var password = ""
+    
+    init() {}
+    
+    func login() {
+        guard !email.trimmingCharacters(in: .whitespaces).isEmpty,
+              !password.trimmingCharacters(in: .whitespaces).isEmpty else {
+            return
+        }
+        
+        print("called")
+    }
+}
