+++ Sweat Social/Views/ContentView.swift
@@ -9,9 +9,8 @@ import SwiftUI
 
 struct ContentView: View {
     var body: some View {
-        NavigationView {
-            LoginView()
-        }
+        LoginView()
+        
         
     }
 }
